Back-end Development/block-23-node.js-service-layer-and-rest/day-04-software-architecture-testing-the-layers/movie-api/services/MoviesService.ts
import { IMoviePayload } from '../compiler/types';
import MoviesModel from '../models/MoviesModel';

function validateMoviePayload(
  title: string,
  directedBy: string,
  releaseYear: number
) {
  if (!title || !directedBy || !releaseYear) return false;

  return true;
}

async function create({ title, directedBy, releaseYear }: IMoviePayload) {
  const isValidMovie = validateMoviePayload(title, directedBy, releaseYear);
  if (!isValidMovie) return false;

  const { id } = await MoviesModel.create({ title, directedBy, releaseYear });

  return id;
}

export default { create };
